pyproject.toml
[tool.poetry]
name = "tech-blog"
version = "0.1.0"
description = ""
authors = ["Abraao Andrade <abraao@nre.tec.br>"]
readme = "README.md"
packages = [{include = "tech_blog"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
pre-commit = "^3.1.1"
django-jazzmin = "^2.6.0"
psycopg2 = "^2.9.5"
dj-database-url = "^1.2.0"
python-decouple = "^3.8"
django-hosts = "^5.2"
drf-yasg = "^1.21.5"
django-extensions = "^3.2.1"
django-cors-headers = "^3.14.0"
django-model-utils = "^4.3.1"
django-ipware = "^5.0.0"
pyjwt = "^2.6.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
redis = "^4.5.1"
celery = "^5.2.7"
pillow = "^9.4.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
